Attempt to make grub actually work
the grub module won't boot with encrypted boot and impermanence
it does boot with unencrypted boot though so I will just replace grub
with systemd-boot for uefi systems since if I can't have encrypted boot
there's no point in using grub. However for bios systems I must use
grub so this is an attempt to get it working with unencrypted boot for
bios systems with impermanence.