Fix sleep during boot + make system gaming capable:
My system was going to sleep during boot if my laptop was docked with
the lid closed because of the switch from i915 to nvidia.
Also opengl and 3D rendering was completely broken because I was using
the wrong openGL driver for my iGPU and I hadn't configured the nvidia
module properly. Now this system can render games with either the nvidia
or intel gpu. Also wine works